Learning management system Shell
Mysql queries and connection already configured.
Current files and images added to folders. Thus far system has login and
registration capability, bootstrap for layout and design not yet
configured. navigation bar is programmed so that when the user is logged
in, the menu displays only logout and when logged out changes to login
and register.