Creating and attempting Quiz
Quiz was created manually and connected to the database with a results
page. A multiple choice question was created allowing the user to select
one answer and advance to a next question using a next button. When the
user has completed all the questions a results page appears upon
clicking the submit button. Users can attempt a question many times
which will be changed. The data from total correct is not yet set up to
be inserted into the database.