Quiz question database cnnection
Connected the quiz to the database and every quiz attempt is now
registered when quiz is clicked in details page. Responses for the quiz
was setup. Response table create in the database to collect the
resnponses to the quiz and store that information in a response variable
which will be checked, if the response is correct then it will be stored
until the quiz is completed then total correct responses will be upated
in the table.