Admin user create topics and Details page continued
Admin added capability to create topics in each course. The admin user
is now capabale of assigning topics to the desired course. The details
page was also modified in the process and some changes are made to
create a table effect which shows a section for lessons and a section
for quizzzes. Failing attempts have been made in order to get the
question type for the pecific enrolled topics.