Admin embedding and uploading files
Page to embed and upload files to the database as well as to display
embeded files. Each embeded file has an (elseif statement) allowing for
different views for different file types. Upload section can only upload
files however files show up as file path, this should be updataed for
files to be used and modified.