Admin Dashboard
Admin is able to do the following on dashboard page without hassle:

Add course to the school
add topic to the course
add lesson to topic
add quiz to topic
add question to topic

user can select whether to embed a lesson or upload a lesson. when
embeded is active upload is disabled i=using jquery. vice versa